requirements.txt
openai-whisper==20231117
moviepy==1.0.3
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.65.0
flask>=2.0.0
